pyproject.toml
[project]
name = "test-deep-research-agents"
version = "0.1.0"
description = "Playground for testing open-source deep-research agents on open benchmarks (SealQA, etc.)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "datasets>=2.19.0",
  "pyarrow>=14.0.1",
  "pandas>=2.2.2",
  "python-dotenv>=1.0.1",
  "tavily-python>=0.7.12",
]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
  "src/test_deep_research_agents",
]

[tool.hatch.metadata]
allow-direct-references = true


